Extract workspace parallel worker pool into microcrate
